Split Application/GraphicsManager/GameClient logic
